+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Plugin changelog
 0.1
 -------
 
+* Added compress option
 * Added lock for share folder
 * Added "addSymbol" task to add symbols to a store.